+++ stroom-processor/stroom-processor-impl/src/main/java/stroom/processor/impl/ProcessorTaskManagerImpl.java
@@ -17,6 +17,7 @@
 
 package stroom.processor.impl;
 
+import stroom.cluster.lock.api.ClusterLockService;
 import stroom.docref.DocRef;
 import stroom.entity.shared.ExpressionCriteria;
 import stroom.meta.api.MetaService;
@@ -107,6 +108,7 @@ class ProcessorTaskManagerImpl implements ProcessorTaskManager {
     private final MetaService metaService;
     private final EventSearch eventSearch;
     private final SecurityContext securityContext;
+    private final ClusterLockService clusterLockService;
 
     private final TaskStatusTraceLog taskStatusTraceLog = new TaskStatusTraceLog();
 
@@ -157,7 +159,8 @@ class ProcessorTaskManagerImpl implements ProcessorTaskManager {
                              final Provider<InternalStatisticsReceiver> internalStatisticsReceiverProvider,
                              final MetaService metaService,
                              final EventSearch eventSearch,
-                             final SecurityContext securityContext) {
+                             final SecurityContext securityContext,
+                             final ClusterLockService clusterLockService) {
 
         this.processorFilterService = processorFilterService;
         this.processorFilterTrackerDao = processorFilterTrackerDao;
@@ -170,6 +173,7 @@ class ProcessorTaskManagerImpl implements ProcessorTaskManager {
         this.metaService = metaService;
         this.eventSearch = eventSearch;
         this.securityContext = securityContext;
+        this.clusterLockService = clusterLockService;
     }
 
     @Override
@@ -268,6 +272,9 @@ class ProcessorTaskManagerImpl implements ProcessorTaskManager {
         // Output some trace logging so we can see where tasks go.
         taskStatusTraceLog.assignTasks(ProcessorTaskManagerImpl.class, assignedStreamTasks, nodeName);
 
+        if (LOGGER.isDebugEnabled()) {
+            LOGGER.debug("Assigning " + assignedStreamTasks.size() + " to node " + nodeName);
+        }
         return new ProcessorTaskList(nodeName, assignedStreamTasks);
     }
 
@@ -293,7 +300,12 @@ class ProcessorTaskManagerImpl implements ProcessorTaskManager {
     private void abandon(final ProcessorTask processorTask) {
         try {
             LOGGER.warn("abandon() - {}", processorTask);
-            processorTaskDao.changeTaskStatus(processorTask, null, TaskStatus.UNPROCESSED, null, null);
+            processorTaskDao.changeTaskStatus(
+                    processorTask,
+                    null,
+                    TaskStatus.UNPROCESSED,
+                    null,
+                    null);
         } catch (final RuntimeException e) {
             LOGGER.error("abandon() - {}", processorTask, e);
         }
@@ -302,7 +314,12 @@ class ProcessorTaskManagerImpl implements ProcessorTaskManager {
     private void release(final ProcessorTask streamTask) {
         try {
             LOGGER.warn("release() - {}", streamTask);
-            processorTaskDao.changeTaskStatus(streamTask, null, TaskStatus.UNPROCESSED, null, null);
+            processorTaskDao.changeTaskStatus(
+                    streamTask,
+                    null,
+                    TaskStatus.UNPROCESSED,
+                    null,
+                    null);
         } catch (final RuntimeException e) {
             LOGGER.error("release() - {}", streamTask, e);
         }
@@ -353,9 +370,12 @@ class ProcessorTaskManagerImpl implements ProcessorTaskManager {
                     if (isScheduled()) {
                         LOGGER.debug("fillTaskStore() - Executing CreateStreamTasksTask");
 
-                        final Runnable runnable = taskContextFactory.context("Fill Task Store", taskContext ->
-                                securityContext.secure(() ->
-                                        createTasks(taskContext)));
+                        final Runnable runnable = taskContextFactory.context(
+                                "Fill Task Store",
+                                taskContext ->
+                                        securityContext.secure(() ->
+                                                createTasks(taskContext)));
+
                         final Executor executor = executorProvider.get(THREAD_POOL);
                         CompletableFuture
                                 .runAsync(runnable, executor)
@@ -409,11 +429,21 @@ class ProcessorTaskManagerImpl implements ProcessorTaskManager {
         // We need to make sure that only 1 thread at a time is allowed to
         // create tasks. This should always be the case in production but some
         // tests will call this directly while scheduled execution could also be
-        // running.
+        // running. Also if the master node changes it is possible for one master
+        // to be in the middle of creating tasks when another node assumes master
+        // status and tries to create tasks too. Thus a db backed cluster lock
+        // is needed
+        // The local lock is still needed to
         createTasksLock.lock();
         try {
             if (allowTaskCreation) {
-                doCreateTasks(taskContext);
+
+                // We need an overarching cluster lock for all task creation but within
+                // doCreateTasks async code will run in other threads that will have to obtain
+                // their own cluster locks using a different lock key
+                clusterLockService.lock("ProcessorTaskManagerImpl.createTasks", () -> {
+                    doCreateTasks(taskContext);
+                });
             }
         } catch (final RuntimeException e) {
             LOGGER.error(e.getMessage(), e);
@@ -455,32 +485,47 @@ class ProcessorTaskManagerImpl implements ProcessorTaskManager {
         // Update the stream task store.
         prioritisedFiltersRef.set(filters);
 
-        // Now fill the stream task store with tasks for each filter.
-        final int totalQueueSize = processorConfig.getQueueSize();
-        final int halfQueueSize = totalQueueSize / 2;
-
         final String nodeName = nodeInfo.getThisNodeName();
         if (nodeName == null) {
             throw new NullPointerException("Node is null");
         }
 
+        // Now fill the stream task store with tasks for each filter.
+        // The aim is to create N tasks in total where N is processorConfig.getQueueSize
+        // Also need to ensure each filter queue has no more than N in it.
+        final int maxTasksToCreate = processorConfig.getQueueSize();
+        final int maxTasksPerFilterQueue = maxTasksToCreate;
+        // If a queue is already half full then don't bother adding more
+        final int halfQueueSize = maxTasksPerFilterQueue / 2;
+
+        final TaskCreationProgressTracker progressTracker = new TaskCreationProgressTracker(
+                maxTasksToCreate,
+                filters.size(),
+                queueMap,
+                processorConfig);
+
         try {
-            int remaining = totalQueueSize;
             for (final ProcessorFilter filter : filters) {
                 final StreamTaskQueue queue = queueMap.computeIfAbsent(filter, k -> new StreamTaskQueue());
-                final int queueSize = queue.size();
+                final int currQueueSize = queue.size();
 
-                // Reduce the number of tasks we need to get by the size of
-                // the
-                // current queue.
-                remaining -= queueSize;
-
-                // Now go and fill this queue asynchronously.
-                if (remaining > 0 && queueSize < halfQueueSize) {
+                if (progressTracker.areTasksRemainingToBeCreated() && currQueueSize < halfQueueSize) {
                     if (queue.compareAndSetFilling(false, true)) {
-                        // Create tasks for this filter.
-                        createTasksForFilter(taskContext, nodeName, filter, queue, totalQueueSize);
+
+                        // Create tasks for this filter
+                        createTasksForFilter(
+                                taskContext,
+                                nodeName,
+                                filter,
+                                queue,
+                                progressTracker);
+                    } else {
+                        // Finished processing this filter
+                        progressTracker.countDownProcessedFilters();
                     }
+                } else {
+                    // Finished processing this filter
+                    progressTracker.countDownProcessedFilters();
                 }
             }
         } catch (final RuntimeException e) {
@@ -510,18 +555,32 @@ class ProcessorTaskManagerImpl implements ProcessorTaskManager {
 //        // Set the last stream details for the next call to this method.
 //        processorTaskManagerRecentStreamDetails = recentStreamInfo;
 
+        // We may have async search tasks still being created so we need to wait for those
+        // in case another node gets master status and tries to do task creation.
+        progressTracker.waitForCompletionOfAllFilters();
+
         taskContext.info(() -> "Finished");
-        LOGGER.debug("doCreateTasks() - Finished in {}", logExecutionTime);
+
+        LOGGER.info(() -> LogUtil.message("Finished creating tasks in {}. {}",
+                logExecutionTime, progressTracker.getProgressSummaryMessage()));
+    }
+
+
+    private int getTaskCountToCreate(final StreamTaskQueue queue,
+                                     final TaskCreationProgressTracker progressTracker) {
+        return Math.min(
+                processorConfig.getQueueSize() - queue.size(),
+                progressTracker.getTotalRemainingTasksToCreate());
     }
 
     private void createTasksForFilter(final TaskContext taskContext,
                                       final String nodeName,
                                       final ProcessorFilter filter,
                                       final StreamTaskQueue queue,
-                                      final int maxQueueSize) {
+                                      final TaskCreationProgressTracker progressTracker) {
         Optional<ProcessorFilter> optionalProcessorFilter = Optional.empty();
 
-        final AtomicBoolean searching = new AtomicBoolean();
+        final AtomicBoolean isSearching = new AtomicBoolean();
         try {
             // Reload as it could have changed
             optionalProcessorFilter = processorFilterService.fetch(filter.getId());
@@ -537,28 +596,32 @@ class ProcessorTaskManagerImpl implements ProcessorTaskManager {
                     if (loadedFilter.isEnabled() && loadedFilter.getProcessor().isEnabled()) {
                         taskContext.info(() -> "Creating tasks: " + loadedFilter.toString());
 
-                        int tasksToCreate = maxQueueSize - queue.size();
-                        int count = 0;
-
                         // If there are any tasks for this filter that were
                         // previously created but are unprocessed, not owned by any
                         // node and their associated stream is unlocked then add
                         // them here.
                         if (processorConfig.isFillTaskQueue()) {
-                            count = addUnownedTasks(taskContext, nodeName, loadedFilter, queue, tasksToCreate);
+                            addUnownedTasks(
+                                    taskContext,
+                                    nodeName,
+                                    loadedFilter,
+                                    queue,
+                                    progressTracker);
                         }
 
                         // If we allowing tasks to be created then go ahead and
                         // create some.
                         if (processorConfig.isCreateTasks()) {
-                            tasksToCreate -= count;
 //                            final String logPrefix = "Creating tasks with filter " + loadedFilter.getId();
 
-                            final Boolean exhausted = exhaustedFilterMap.computeIfAbsent(loadedFilter.getId(),
+                            final Boolean exhausted = exhaustedFilterMap.computeIfAbsent(
+                                    loadedFilter.getId(),
                                     k -> Boolean.FALSE);
 
                             // Skip once we have done all that is required
-                            final int requiredTasks = tasksToCreate;
+                            // re-compute tasks to create after adding unowned tasks
+                            final int requiredTasks = getTaskCountToCreate(queue, progressTracker);
+
                             if (requiredTasks > 0 && !Thread.currentThread().isInterrupted()) {
                                 final QueryData queryData = loadedFilter.getQueryData();
                                 final boolean isStreamStoreSearch = queryData.getDataSource() != null
@@ -590,25 +653,24 @@ class ProcessorTaskManagerImpl implements ProcessorTaskManager {
                                     }
 
                                 } else if (!isStreamStoreSearch) {
-                                    // Create stream tasks by executing a
-                                    // search.
-                                    searching.set(true);
-                                    createTasksFromSearchQuery(loadedFilter,
+                                    // Create stream tasks by executing a search.
+                                    isSearching.set(true);
+                                    createTasksFromSearchQuery(
+                                            loadedFilter,
                                             queryData,
                                             streamQueryTime,
                                             nodeName,
-                                            requiredTasks,
+                                            progressTracker,
                                             queue,
                                             tracker);
-
                                 } else {
-                                    // Create tasks from a standard stream
-                                    // filter criteria.
-                                    createTasksFromCriteria(loadedFilter,
+                                    // Create tasks from a standard stream filter criteria.
+                                    createTasksFromCriteria(
+                                            loadedFilter,
                                             queryData,
                                             streamQueryTime,
                                             nodeName,
-                                            requiredTasks,
+                                            progressTracker,
                                             queue,
                                             tracker);
                                 }
@@ -623,8 +685,8 @@ class ProcessorTaskManagerImpl implements ProcessorTaskManager {
                             exhaustedFilterMap.put(loadedFilter.getId(), Boolean.FALSE);
                         }
 
-                        final long c = count;
-                        taskContext.info(() -> "Created " + c + " tasks: " + loadedFilter.toString());
+                        taskContext.info(() -> "Created " + progressTracker.getCreatedCount(loadedFilter)
+                                + " tasks: " + loadedFilter.toString());
                     }
                 });
             });
@@ -642,8 +704,12 @@ class ProcessorTaskManagerImpl implements ProcessorTaskManager {
             LOGGER.error(() -> "Error processing filter with id = " + filter.getId() + pipelineDetails);
             LOGGER.error(e::getMessage, e);
         } finally {
-            if (!searching.get()) {
+            if (!isSearching.get()) {
                 queue.setFilling(false);
+
+                // Not searching (which is async) so tick off this filter as done
+                // Async search will tick off its filter when it completes
+                progressTracker.countDownProcessedFilters();
             }
         }
     }
@@ -652,8 +718,9 @@ class ProcessorTaskManagerImpl implements ProcessorTaskManager {
                                 final String nodeName,
                                 final ProcessorFilter filter,
                                 final StreamTaskQueue queue,
-                                final int tasksToCreate) {
+                                final TaskCreationProgressTracker progressTracker) {
         int count = 0;
+        int tasksToCreate = getTaskCountToCreate(queue, progressTracker);
 
         try {
             // First look for any items that are no-longer locked etc
@@ -727,6 +794,7 @@ class ProcessorTaskManagerImpl implements ProcessorTaskManager {
             }
 
             if (count > 0) {
+                progressTracker.incrementTaskCreationCount(filter, count);
                 LOGGER.debug("doCreateTasks() - Added {} tasks that are no longer locked", count);
             }
         } catch (final RuntimeException e) {
@@ -740,12 +808,14 @@ class ProcessorTaskManagerImpl implements ProcessorTaskManager {
                                             final QueryData queryData,
                                             final long streamQueryTime,
                                             final String nodeName,
-                                            final int requiredTasks,
+                                            final TaskCreationProgressTracker progressTracker,
                                             final StreamTaskQueue queue,
                                             final ProcessorFilterTracker tracker) {
+
         final EventRef minEvent = new EventRef(tracker.getMinMetaId(), tracker.getMinEventId());
         final EventRef maxEvent = new EventRef(Long.MAX_VALUE, 0L);
-        long maxStreams = requiredTasks;
+        long maxStreams = progressTracker.getTaskCountToCreate(filter);
+        LOGGER.debug("Creating search query tasks maxStreams: {}, filer: {}", maxStreams, filter);
         long maxEvents = 1000000;
         final long maxEventsPerStream = 1000;
 
@@ -814,22 +884,26 @@ class ProcessorTaskManagerImpl implements ProcessorTaskManager {
                 maxEvents,
                 maxEventsPerStream,
                 POLL_INTERVAL_MS,
-                eventRefs ->
-                        createTasksFromEventRefs(
-                                filter,
-                                streamQueryTime,
-                                nodeName,
-                                requiredTasks,
-                                queue,
-                                maxMetaId,
-                                updatedTracker,
-                                eventRefs));
+                eventRefs -> {
+                    createTasksFromEventRefs(
+                            filter,
+                            streamQueryTime,
+                            nodeName,
+                            progressTracker,
+                            queue,
+                            maxMetaId,
+                            updatedTracker,
+                            eventRefs);
+
+                    // We have now completed tasking for this filter so count down our counter
+                    progressTracker.countDownProcessedFilters();
+                });
     }
 
     private void createTasksFromEventRefs(final ProcessorFilter filter,
                                           final long streamQueryTime,
                                           final String nodeName,
-                                          final int requiredTasks,
+                                          final TaskCreationProgressTracker progressTracker,
                                           final StreamTaskQueue queue,
                                           final Long maxMetaId,
                                           final ProcessorFilterTracker updatedTracker,
@@ -851,6 +925,7 @@ class ProcessorTaskManagerImpl implements ProcessorTaskManager {
 
             // Create a task for each stream reference.
             final Map<Meta, InclusiveRanges> map = createStreamMap(eventRefs);
+
             processorTaskDao.createNewTasks(
                     filter,
                     tracker,
@@ -862,10 +937,15 @@ class ProcessorTaskManagerImpl implements ProcessorTaskManager {
                     createdTasks -> {
                         // Transfer the newly created (and available) tasks to the
                         // queue.
-                        createdTasks.getAvailableTaskList().forEach(queue::add);
-                        LOGGER.debug("createTasks() - Created {} tasks (tasksToCreate={}) for filter {}",
+                        final List<ProcessorTask> availableTaskList = createdTasks.getAvailableTaskList();
+
+                        if (!availableTaskList.isEmpty()) {
+                            queue.addAll(availableTaskList);
+                            progressTracker.incrementTaskCreationCount(filter, availableTaskList.size());
+                        }
+
+                        LOGGER.debug("createTasks() - Created {} tasks for filter {}",
                                 createdTasks.getTotalTasksCreated(),
-                                requiredTasks,
                                 filter.toString());
 
                         exhaustedFilterMap.put(filter.getId(), exhausted);
@@ -879,51 +959,60 @@ class ProcessorTaskManagerImpl implements ProcessorTaskManager {
                                          final QueryData queryData,
                                          final long streamQueryTime,
                                          final String nodeName,
-                                         final int requiredTasks,
+                                         final TaskCreationProgressTracker progressTracker,
                                          final StreamTaskQueue queue,
                                          final ProcessorFilterTracker tracker) {
         if (termCount(queryData) == 0) {
             throw new RuntimeException("Attempting to create tasks with an unconstrained filter " + filter);
         }
 
-        // Update the tracker status message.
-        tracker.setStatus("Creating...");
-        final ProcessorFilterTracker updatedTracker = processorFilterTrackerDao.update(tracker);
+        final int requiredTasks = progressTracker.getTaskCountToCreate(filter);
+        LOGGER.debug("Creating tasks from criteria, requiredTasks: {}, filter: {}", requiredTasks, filter);
 
-        // This will contain locked and unlocked streams
-        final Long maxMetaId = metaService.getMaxId();
-        final List<Meta> streamList = runSelectMetaQuery(
-                queryData.getExpression(),
-                updatedTracker.getMinMetaId(),
-                updatedTracker.getMinMetaCreateMs(),
-                updatedTracker.getMaxMetaCreateMs(),
-                filter.getPipeline(),
-                filter.isReprocess(),
-                requiredTasks);
+        if (requiredTasks > 0) {
+            // Update the tracker status message.
+            tracker.setStatus("Creating...");
 
-        // Just create regular stream processing tasks.
-        final Map<Meta, InclusiveRanges> map = new HashMap<>();
-        for (final Meta meta : streamList) {
-            map.put(meta, null);
+            final ProcessorFilterTracker updatedTracker = processorFilterTrackerDao.update(tracker);
+
+            // This will contain locked and unlocked streams
+            final Long maxMetaId = metaService.getMaxId();
+            final List<Meta> streamList = runSelectMetaQuery(
+                    queryData.getExpression(),
+                    updatedTracker.getMinMetaId(),
+                    updatedTracker.getMinMetaCreateMs(),
+                    updatedTracker.getMaxMetaCreateMs(),
+                    filter.getPipeline(),
+                    filter.isReprocess(),
+                    requiredTasks);
+
+            // Just create regular stream processing tasks.
+            final Map<Meta, InclusiveRanges> map = new HashMap<>();
+            for (final Meta meta : streamList) {
+                map.put(meta, null);
+            }
+
+            processorTaskDao.createNewTasks(
+                    filter,
+                    updatedTracker,
+                    streamQueryTime,
+                    map,
+                    nodeName,
+                    maxMetaId,
+                    false,
+                    createdTasks -> {
+                        // Transfer the newly created (and available) tasks to the queue.
+                        final List<ProcessorTask> availableTaskList = createdTasks.getAvailableTaskList();
+                        queue.addAll(availableTaskList);
+                        progressTracker.incrementTaskCreationCount(filter, availableTaskList.size());
+
+                        LOGGER.debug("createTasks() - Created {} tasks (requiredTasks={}) for filter {}",
+                                createdTasks.getTotalTasksCreated(),
+                                requiredTasks,
+                                filter.toString());
+                        exhaustedFilterMap.put(filter.getId(), createdTasks.getTotalTasksCreated() == 0);
+                    });
         }
-
-        processorTaskDao.createNewTasks(
-                filter,
-                updatedTracker,
-                streamQueryTime,
-                map,
-                nodeName,
-                maxMetaId,
-                false,
-                createdTasks -> {
-                    // Transfer the newly created (and available) tasks to the queue.
-                    createdTasks.getAvailableTaskList().forEach(queue::add);
-                    LOGGER.debug("createTasks() - Created {} tasks (tasksToCreate={}) for filter {}",
-                            createdTasks.getTotalTasksCreated(),
-                            requiredTasks,
-                            filter.toString());
-                    exhaustedFilterMap.put(filter.getId(), createdTasks.getTotalTasksCreated() == 0);
-                });
     }
 
     private int termCount(final QueryData queryData) {
@@ -1063,7 +1152,8 @@ class ProcessorTaskManagerImpl implements ProcessorTaskManager {
     private void scheduleDelete() {
         if (nextDeleteMs.get() == 0) {
             nextDeleteMs.set(System.currentTimeMillis() + DELETE_INTERVAL_MS);
-            LOGGER.debug("scheduleDelete() - nextDeleteMs={}", DateUtil.createNormalDateTimeString(nextDeleteMs.get()));
+            LOGGER.debug("scheduleDelete() - nextDeleteMs={}",
+                    DateUtil.createNormalDateTimeString(nextDeleteMs.get()));
         }
     }
 
@@ -1109,4 +1199,5 @@ class ProcessorTaskManagerImpl implements ProcessorTaskManager {
     public void setAllowTaskCreation(final boolean allowTaskCreation) {
         this.allowTaskCreation = allowTaskCreation;
     }
+
 }